+++ src/auth/register.js
@@ -1,5 +1,3 @@
-import { manageErrors } from '../utils/commons';
-
 export const registerUser = async (userInfo) => {
   const response = await fetch('http://localhost:3000/register', {
     method: 'POST',
@@ -8,7 +6,6 @@ export const registerUser = async (userInfo) => {
       'Content-Type': 'application/json',
     },
   })
-    .then(manageErrors)
     .then((res) => res.json())
     .catch(({ message }) => {
       window.alert(message);
